Move externals handling logic into the loader to remove the need for actual fetch
